+++ CHANGELOG.md
@@ -9,13 +9,18 @@
  how a consumer would use the library or CLI tool (e.g. adding unit tests, updating documentation, etc) are not captured
  here.
 
-## Unreleased
+## 2.12.0 - 2026-05-04
 
 ### Added
 - Added the `sdk.file-events.v2.search_groups` method to get approximate aggregate file event counts by a given grouping term.
 - Added the `GroupingEventQuery` class, used to make these queries.
 - Added the cli command `incydr file-events search-groups` to get approximate aggregate file event counts by a given grouping term.
-
+- Added the `type` parameter to session search methods and commands, allowing users to filter results to STANDARD or ACCOUNT_TAKE_OVER.
+- Added the `is_high_value` option to trusted activity methods in the SDK, and the `--high-value` option to trusted activity methods in the CLI.
+- Added the ability to specify domain trust for browser destinations, allowing users to specify when users should be allowed to use certain destinations when logged-in using a trusted domain.
+- Added the ability to specify trust for file-transfer tools when adding a trusted domain.
+- Added the `risk-indicator-categories` client to the SDK, allowing the listing of risk indicator categories, subcategories, and risk indicators.
+- Added the `risk-indicator-categories` command to the CLI, allowing the listing of risk indicator categories.
 
 ## 2.11.0 - 2026-02-10